fix type hints, pass through loss weight

+++ gliner2/training/data.py
@@ -649,7 +649,7 @@ class InputExample:
     classifications: Optional[List[Classification]] = None
     structures: Optional[List[Structure]] = None
     relations: Optional[List[Relation]] = None
-    loss_weight: int = 1
+    loss_weight: float = 1.0
 
     def __post_init__(self):
         if self.entities is None:
